WIP: v2 rewrite
This simplifies the whole confita API with an aim to making the behavior
more predictable. In particular, the treatment of backends is more consistent;
config fields are consistently used and named.

The exported API is smaller (no need for three ways to specify a backend
unmarshaler, and the backend semantics are defined such that it should
be easy to make concurrent without changing behaviour.