backend/env: treat empty environment variables as not found (#78)
It's very unusual to distinguish between an environment variable
that's unset and one that's empty. When a value is required, looking
for a non-empty environment variable feels more correct.

As a recent example, the kafka-go config contained a "required" entry
for `KAFKA_ADDRS` which was satisfied by an empty environment
variable.