backend/flags: avoiding setting values that aren't specified
In #67, the `isFlagSet` logic was changed so that it always returned
true, so even unmentioned flags would override previously set values.

This PR fixes that issue (fixes #79) and also changes the flag tests
to be somewhat more flexible and test this specific issue.

I haven't made any wider fixes because the concept of using command-line
flags as a backend is fundamentally broken in a way that cannot be fixed
(see #63), so will probably be removed in a future major version release
of confita.